fix(minimap): preserve scoped navigation order
